CAS-488: message bus timeout and retry
Message bus requests can go unanswered. In fact, if no receiver was
listening in when the request was sent, it will not get it altogether.

This a default timeout of 5s with a default max number of 5 retries.
To avoid complicating things for the moment, each retry it spaced out by
 1s * nr of retries and capped at 10s - further down the line we might
want to customize this a bit.

By using request_ext one can specify timeout and max_retries explicitly.